// web/src/api/models/CreateExternalIdentityMappingRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateExternalIdentityMappingRequest = {
  external_subject: string;
  mapped_identity: number;
  provider: string;
  subject_kind: string;
  tenant: string;
};